fix: missing icon on taskbar
